Use PromptTemplate for Pytest snippets
