+++ lavague-qa/lavague/qa/prompts.py
@@ -1,5 +1,58 @@
 from llama_index.core import PromptTemplate
 
+
+PYTEST_HEADER_TEMPLATE = PromptTemplate("""
+import pytest
+from pytest_bdd import scenarios, given, when, then
+from selenium.webdriver.chrome.webdriver import WebDriver
+from selenium.webdriver.common.by import By
+from selenium.webdriver.support.ui import WebDriverWait, Select
+from selenium.webdriver.support import expected_conditions as EC
+
+# Constants
+BASE_URL = '{url}'
+
+# Scenarios
+scenarios('{feature_file_name}')
+
+# Fixtures
+@pytest.fixture
+def browser():
+    driver = WebDriver()
+    driver.implicitly_wait(10)
+    yield driver
+    driver.quit()
+
+# Steps
+""")
+
+PYTEST_GIVEN_TEMPLATE = PromptTemplate("""@given('{step}')
+def {method_name}(browser: WebDriver):
+    {code}
+"""
+)
+
+PYTEST_WHEN_TEMPLATE = PromptTemplate("""@when('{step}')
+def {method_name}(browser: WebDriver):
+{actions_code}
+    """)
+
+PYTEST_THEN_TEMPLATE = PromptTemplate("""@then('{step}')
+def {method_name}(browser: WebDriver):
+{assert_code}
+""")
+
+
+
+
+
+
+
+
+
+
+
+
 FULL_PROMPT_TEMPLATE = PromptTemplate("""You are an expert in software testing frameworks and Python code generation. You answer in python markdown only and nothing else.
 Your only goal is to generate pytest-bdd files based on the provided Gherkin feature, a collection of instructions and actions, and a specific assert statement to test.
 You will use the provided information to generate a valid assert statement.
